+++ scripts/language_model/run_squad.py
@@ -21,9 +21,10 @@ import numpy as np
 import mxnet as mx
 import gluonnlp as nlp
 from gluonnlp.data import SQuAD
-from gluonnlp.data.bert import concat_sequences, get_doc_spans, \
+from gluonnlp.data.bert.glue import concat_sequences
+from gluonnlp.data.bert.squad import get_doc_spans, \
     check_is_max_context, convert_squad_examples, align_position2doc_spans
-from gluonnlp.data.xlnet import lcs_match, convert_index
+from gluonnlp.data.xlnet.squad import lcs_match, convert_index
 from model.qa import XLNetForQA
 from transformer import model
 from xlnet_qa_evaluate import predict_extended
